backend/ai_service/requirements.txt
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
pydantic>=2.7.0
pypdf>=4.0.0
requests>=2.31.0
numpy>=1.24.0
sentence-transformers>=2.7.0
fastapi>=0.115.0
uvicorn>=0.30.0